requirements.txt
python-dotenv
fastapi 
uvicorn 
pydantic
mysql-connector-python
bcrypt

